マージミスを大体直した クリスタルブロックだけだめかも
