+++ src/main/resources/assets/fantalmod/lang/ja_jp.json
@@ -13,5 +13,11 @@
     "item.fantalmod.fantal_sword" : "ファンタルの剣",
     "item.fantalmod.fantal_axe" : "ファンタルの斧",
     "item.fantalmod.fantal_shovel" : "ファンタルのシャベル",
-    "item.fantalmod.fantal_hoe" : "ファンタルのクワ"
+    "item.fantalmod.fantal_hoe" : "ファンタルのクワ",
+    "item.modid.test_item" : "テストアイテム",
+    "item.modid.anime_item" : "アニメアイテム",
+    "block.modid.test_block" : "テストブロック",
+    "block.modid.anime_block" : "アニメブロック",
+    "block.modid.crystal_block" : "クリスタルブロック",
+    "block.modid.clear_block" : "クリアブロック"
 }
