+++ src/main/resources/assets/fantalmod/models/block/anime_block.json
@@ -1,6 +1,6 @@
 {
   "parent" : "block/cube_all",
   "textures" : {
-    "all" : "modid:block/anime_block"
+    "all" : "fantalmod:block/anime_block"
   }
 }
