+++ src/main/resources/assets/fantalmod/models/block/clear_block.json
@@ -1,6 +1,6 @@
 {
   "parent" : "block/cube_all",
   "textures" : {
-    "all" : "modid:block/clear_block"
+    "all" : "fantalmod:block/clear_block"
   }
 }
