+++ src/main/resources/assets/fantalmod/models/block/crystal_block.json
@@ -1,6 +1,6 @@
 {
   "parent" : "block/cross",
   "textures" : {
-    "cross" : "modid:block/crystal_block"
+    "cross" : "fantalmod:block/crystal_block"
   }
 }
