+++ src/main/resources/assets/fantalmod/models/item/anime_block.json
@@ -1,3 +1,3 @@
 {
-  "parent" : "modid:block/anime_block"
+  "parent" : "fantalmod:block/anime_block"
 }
