+++ src/main/resources/assets/fantalmod/models/item/anime_item.json
@@ -1,6 +1,6 @@
 {
   "parent" : "item/generated",
   "textures" : {
-    "layer0" : "modid:item/anime_item"
+    "layer0" : "fantalmod:item/anime_item"
   }
 }
