+++ src/main/resources/assets/fantalmod/models/item/clear_block.json
@@ -1,3 +1,3 @@
 {
-    "parent" : "modid:block/clear_block"
+    "parent" : "fantalmod:block/clear_block"
 }
