+++ src/main/resources/assets/fantalmod/textures/block/anime_block.png.mcmeta
