+++ src/main/resources/assets/fantalmod/textures/block/clear_block.png
