+++ src/main/resources/assets/fantalmod/textures/block/crystal_block.png
