--- src/main/resources/assets/modid/lang/en_us.json
@@ -1,13 +1,0 @@
-{
-    "item.modid.test_item" : "Test Item",
-    "item.modid.anime_item" : "Anime Item",
-    "item.modid.tutorial_pickaxe" : "Tutorial Pickaxe",
-    "item.modid.tutorial_helmet" : "Tutorial Helmet",
-    "item.modid.tutorial_chestplate" : "Tutorial ChestPlate",
-    "item.modid.tutorial_leggings" : "Tutorial Leggings",
-    "item.modid.tutorial_boots" : "Tutorial Boots",
-    "block.modid.test_block" : "Test Block",
-    "block.modid.anime_block" : "Anime Block",
-    "block.modid.crystal_block" : "Crystal Block",
-    "block.modid.clear_block" : "clear_block"
-}
